refactor: export tx schemas from a dedicated file
refs #1272
