+++ apps/api/src/billing/controllers/wallet/wallet.controller.ts
@@ -4,13 +4,8 @@ import { Lifecycle, scoped } from "tsyringe";
 
 import { AuthService, Protected } from "@src/auth/services/auth.service";
 import type { GetBalancesResponseOutput } from "@src/billing/http-schemas/balance.schema";
-import type {
-  SignTxRequestInput,
-  SignTxResponseOutput,
-  StartTrialRequestInput,
-  WalletListOutputResponse,
-  WalletOutputResponse
-} from "@src/billing/http-schemas/wallet.schema";
+import type { SignTxRequestInput, SignTxResponseOutput } from "@src/billing/http-schemas/tx.schema";
+import type { StartTrialRequestInput, WalletListOutputResponse, WalletOutputResponse } from "@src/billing/http-schemas/wallet.schema";
 import { UserWalletRepository } from "@src/billing/repositories";
 import type { GetWalletQuery } from "@src/billing/routes/get-wallet-list/get-wallet-list.router";
 import { WalletInitializerService } from "@src/billing/services";
